feat: move SQL n:m extraction tests to dedicated files
